Day 25, no external libraries, simple solution.
Solution using Karger's algorithm to partition the graph into two, using
the fact that we know the min cut is of order 3.

Day 25, no external libraries, simple solution.

Solution using Karger's algorithm to partition the graph into two, using
the fact that we know the min cut is of order 3.