More rigorous day 24 part 2
The solution is still funky. For the actual input even with the correct
velocity adjusting the hailstones according to that velocity and then
finding the intersection between hailstones doesn't all give the same
intersection. That is the error is never 0 (it is for the test input).
This might be because of floating point arithmetic?